+++ packages/session-cookie/src/index.ts
@@ -1,4 +1,5 @@
-export { SessionCookieOptions } from './types.js';
+export { SessionCookieOptions, RequestContextWithSession } from './types.js';
 export { SessionCookie } from './session-cookie.js';
 export { SessionCookieModule } from './session-cookie.module.js';
 export { SessionLogMediator } from './session-log-mediator.js';
+export { SessionCookieExtension, SESSION_COOKIE_EXTENSIONS } from './session-cookie.extension.js';