+++ packages/session-cookie/src/session-cookie.module.ts
@@ -1,13 +1,23 @@
 import { featureModule, ModuleWithParams, optional } from '@ditsmod/core';
+import { PRE_ROUTER_EXTENSIONS } from '@ditsmod/routing';
 
 import { SessionCookie } from './session-cookie.js';
 import { SessionLogMediator } from './session-log-mediator.js';
 import { SessionCookieOptions } from './types.js';
+import { SESSION_COOKIE_EXTENSIONS, SessionCookieExtension } from './session-cookie.extension.js';
 
 @featureModule({
   providersPerMod: [SessionLogMediator],
   providersPerReq: [SessionCookie],
   exports: [SessionCookie],
+  extensions: [
+    {
+      extension: SessionCookieExtension,
+      groupToken: SESSION_COOKIE_EXTENSIONS,
+      nextToken: PRE_ROUTER_EXTENSIONS,
+      exported: true,
+    },
+  ],
 })
 export class SessionCookieModule {
   static withParams(opts: SessionCookieOptions): ModuleWithParams<SessionCookieModule> {