+++ packages/session-cookie/CHANGELOG.md
@@ -1,5 +1,12 @@
+<a name="session-cookie-2.7.0"></a>
+# [session-cookie-2.7.0](https://github.com/ditsmod/ditsmod/releases/tag/session-cookie-2.7.0) (2023-10-12)
+
+### Features
+
+- Added support for controllers singletons (see example 19).
+
 <a name="session-cookie-2.6.0"></a>
-# [session-cookie-2.6.0](https://github.com/ditsmod/ditsmod/releases/tag/session-cookie-2.6.0) (2023-08-28)
+## [session-cookie-2.6.0](https://github.com/ditsmod/ditsmod/releases/tag/session-cookie-2.6.0) (2023-08-28)
 
 ### Features and Breaking changes